requirements.txt
Flask==1.1.2
python-dotenv==0.13.0
pyTelegramBotAPI==3.7.1
psycopg2-binary==2.8.5
